\name{GOrd}
\alias{GOrd}
\title{Genomic additive relationship matrix for the GLS dataset.}
\description{
Genomic relationship matrix for the GLS dataset. 
The matrix was derived from 46347 markers for the 278 individuals. 
The matrix was calculated as follows G=MM'/p, where
M is the matrix of markers centered and standarized and p is the 
number of markers. 
}

\source{
  International Maize and Wheat Improvement Center (CIMMYT), Mexico.
}

\references{
Montesinos-Lopez, O., A. Montesinos-Lopez, J. Crossa, J. Burgueno and K. Eskridge. 2015. "Genomic-Enabled Prediction of Ordinal Data with Bayesian Logistic Ordinal Regression". \emph{G3: Genes | Genomes | Genetics}, \bold{5}, 2113-2126.
}

\keyword{datasets}
