\name{partitions}
\alias{partitions}
\title{Partitions for cross validation (CV)}
\description{
Is a vector \eqn{(297 \times 1)}{(297 x 1)} that assigns observations to 10
disjoint sets; the assignment was generated at random.
This is used later to conduct a 10-fold CV.
}

\source{
  University of Wisconsin at Madison, USA.
}

\keyword{datasets}
