\name{pheno}
\alias{pheno}
\title{Phenotypic information for Jersey}
\description{ 
The format of the phenotype dataframe is animal ID, herd, 
number of lactations, average milk yield, average fat yield, 
average protein yield, yield deviation for milk, 
yield deviation for fat, and yield deviation for protein.  
Averages are adjusted for age, days in milk, and milking frequency.  
Yield deviations are adjusted further, for herd-year-season effect.  
You may wish to use yield deviations, because there are relatively 
few cows per herd (farmers don't pay to genotype all of their cows, just the good ones).
}

\references{
Gianola, D. Okut, H., Weigel, K. and Rosa, G. 2011. "Predicting complex quantitative traits with Bayesian neural networks: a case study with Jersey cows and wheat". \emph{BMC Genetics}, 
\bold{12},87.
}

\source{
  University of Wisconsin at Madison, USA.
}

\keyword{datasets}
