\name{phenoOrd}
\alias{phenoOrd}
\title{Phenotypic information for GLS (ordinal trait)}
\description{ 
The format of the phenotype dataframe is Location (Loc), 
Replicate (Rep), Genotype ID (Stock) and rating (ordinal score).
Gray Leaf Spot (GLS) is a disease caused by the fungus 
\emph{Cercospora zeae-maydis}. This dataset consists of 
genotypic and phenotypic information for 278 maize lines from 
the Drought Tolerance Maize (DTMA) project of CIMMYT's Global 
Maize Program. The dataset includes information on disease 
severity measured on an ordinal scale with 5 points: 
1= no disease, 2= low infection, 3=moderate infection, 
4=high infection and 5=totally infected.
}

\references{
Montesinos-Lopez, O., A. Montesinos-Lopez, J. Crossa, J. Burgueno and K. Eskridge. 2015. "Genomic-Enabled Prediction of Ordinal Data with Bayesian Logistic Ordinal Regression". \emph{G3: Genes | Genomes | Genetics}, \bold{5}, 2113-2126.
}

\source{
  International Maize and Wheat Improvement Center (CIMMYT), Mexico.
}

\keyword{datasets}
