\name{predictions.nn}
\alias{predictions.nn}
\title{predictions.nn}
\description{
  The function produces the predictions for a two-layer feed-forward neural network.
}
\usage{
   predictions.nn(X,theta,neurons)
}
\arguments{
  \item{X}{X is the matrix with data.}
  \item{theta}{theta is a list with the network parameters (weights and biases).}
  \item{neurons}{neurons is the number of neurons.}
}

\details{
  The details
}
\value{
  A vector containig the predictions.
}



\keyword{models}
