\name{twoinput}
\alias{twoinput}
\title{2 Inputs and 1 output.}
\description{

The data used in Paciorek and
Schervish (2004). This is a data.frame with 3 columns, columns 1 and 2 corresponds to the 
predictors and column 3 corresponds to the target.

}

\references{

Paciorek, C. J. and Schervish, M. J. 2004. "Nonstationary Covariance Functions for
Gaussian Process Regression". In Thrun, S., Saul, L., and Scholkopf, B., editors, \emph{Advances
in Neural Information Processing Systems 16}. MIT Press, Cambridge, MA.

}
\keyword{datasets}
